@@ -4,6 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
+    "rxjs": "rxjs-5-to-6-migrate -p src/tsconfig.app.json",
     "start": "webpack-dev-server --port 2992 --content-base build/public --hot --inline",
     "build": "webpack --progress --profile --colors",
     "production-build": "webpack --progress --profile --colors --env.production --env.DEPLOY_URL='./assets/'",
@@ -14,38 +15,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@types/file-saver": "^1.3.0",
     "@zxing/library": "^0.12.3",
-    "bootstrap": "^4.3.1",
     "core-js": "^2.4.1",
     "file-saver": "^1.3.8",
     "font-awesome": "^4.7.0",
-    "jquery": "^3.4.0",
-    "named-regexp-groups": "^1.0.3",
     "ng2-file-upload": "^1.2.1",
     "ngx-bootstrap": "^1.7.1",
     "popper.js": "^1.12.9",
-    "rxjs": "^5.1.0",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.0.0-rc.0",
+    "rxjs-tslint": "^0.1.7",
     "tachyons": "^4.9.0",
-    "zone.js": "^0.8.4"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "1.1.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/language-service": "^4.0.0",
+    "@angular-devkit/build-angular": "~0.803.20",
+    "@angular/cli": "^8.3.20",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "2.5.45",
     "@types/node": "~6.0.60",
     "autoprefixer": "^6.5.3",
-    "codelyzer": "~3.0.1",
+    "codelyzer": "^5.0.1",
     "css-loader": "^0.28.1",
     "cssnano": "^3.10.0",
     "exports-loader": "^0.6.3",
@@ -61,6 +62,7 @@
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "less-loader": "^4.0.2",
+    "less": "^2.3.1",
     "manifest-revision-webpack-plugin": "^0.4.0",
     "postcss-loader": "^1.3.3",
     "postcss-url": "^5.1.2",
@@ -74,9 +76,9 @@
     "sync-exec": "^0.6.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3",
+    "typescript": "~3.5.3",
     "url-loader": "^0.5.7",
-    "webpack": "~2.4.0",
-    "webpack-dev-server": "~2.4.5"
+    "webpack": "^4.6.0",
+    "webpack-dev-server": "^3.1.0"
   }
 }
