@@ -1,7 +1,7 @@
 import { AppRoutingModule } from './app-routing.module';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule, LOCALE_ID } from '@angular/core';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { ReactiveFormsModule } from '@angular/forms';
 
 import { SharedModule } from './shared/shared.module';
@@ -113,7 +113,7 @@ import { AppComponent } from './app.component';
         AttributeDefinitionTitlePipe
     ],
     imports: [
-        HttpModule,
+        HttpClientModule,
         BrowserModule,
         ReactiveFormsModule,
         SharedModule,
