@@ -1,8 +1,10 @@
+
+import {take} from 'rxjs/operators';
 import { Component, OnChanges, Input, ViewChild, OnDestroy } from '@angular/core';
 import { Router } from '@angular/router';
 import { ApiObject } from '../shared/rest/api-base.service';
 import { ApiService } from '../shared/rest/api.service';
-import { Subscription } from 'rxjs/Rx';
+import { Subscription } from 'rxjs';
 import { DynamicFormComponent } from '../shared/forms/dynamic-form/dynamic-form.component';
 import { JWTService } from '../shared/rest/jwt.service';
 import { QuestionBase } from 'app/shared/forms/question-base';
@@ -18,7 +20,7 @@ export class AttributeDefinitionEditComponent implements OnChanges, OnDestroy {
 
     private subscription: Subscription;
 
-    @ViewChild(DynamicFormComponent) form;
+    @ViewChild(DynamicFormComponent, { static: true }) form;
 
     @Input() attributeDefinitionID: number;
 
@@ -69,7 +71,7 @@ export class AttributeDefinitionEditComponent implements OnChanges, OnDestroy {
             properties: {
                 [attribute_definition.name]: schema,
             }
-        }).take(1).subscribe(questions => {
+        }).pipe(take(1)).subscribe(questions => {
             this.questions = questions;
             this.questions.forEach(qstn => {
                 if (qstn.key === attribute_definition.name) {
@@ -91,7 +93,7 @@ export class AttributeDefinitionEditComponent implements OnChanges, OnDestroy {
                 event.jsonschema = JSON.stringify(JSON.parse(event.jsonschema), undefined, '\t');
             } catch (error) {}
         }
-        this.api.putAttributeDefinition(this.attrDef.id, event).take(1).subscribe(() => {
+        this.api.putAttributeDefinition(this.attrDef.id, event).pipe(take(1)).subscribe(() => {
             this.form.saveFinished(true);
             this.updateTestForm();
         }, () => {
@@ -100,7 +102,7 @@ export class AttributeDefinitionEditComponent implements OnChanges, OnDestroy {
     }
 
     delete = () => {
-        this.api.deleteAttributeDefinition(this.attrDef.id).take(1).subscribe(() => this.router.navigate(['attribute-definitions']));
+        this.api.deleteAttributeDefinition(this.attrDef.id).pipe(take(1)).subscribe(() => this.router.navigate(['attribute-definitions']));
     };
 
 }
