@@ -1,8 +1,10 @@
+
+import {take} from 'rxjs/operators';
 import { Component, OnChanges, Input, OnDestroy, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import { ApiObject } from '../shared/rest/api-base.service';
 import { ApiService } from '../shared/rest/api.service';
-import { Subscription } from 'rxjs/Rx';
+import { Subscription } from 'rxjs';
 import { DynamicFormComponent } from '../shared/forms/dynamic-form/dynamic-form.component';
 import { JWTService } from '../shared/rest/jwt.service';
 
@@ -14,7 +16,7 @@ export class TagEditComponent implements OnChanges, OnDestroy {
 
     private subscription: Subscription;
 
-    @ViewChild(DynamicFormComponent) form;
+    @ViewChild(DynamicFormComponent, { static: true }) form;
 
     @Input() tagID: number;
 
@@ -41,7 +43,7 @@ export class TagEditComponent implements OnChanges, OnDestroy {
     }
 
     save = (event) => {
-        this.api.putTag(this.tag.id, event).take(1).subscribe(() => {
+        this.api.putTag(this.tag.id, event).pipe(take(1)).subscribe(() => {
             this.form.saveFinished(true);
         }, () => {
             this.form.saveFinished(false);
@@ -49,7 +51,7 @@ export class TagEditComponent implements OnChanges, OnDestroy {
     }
 
     delete = () => {
-        this.api.deleteTag(this.tag.id).take(1).subscribe(() => this.router.navigate(['tags']));
+        this.api.deleteTag(this.tag.id).pipe(take(1)).subscribe(() => this.router.navigate(['tags']));
     };
 
 }
